pyproject.toml
[project]
name = "yoyoai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.68.0",
    "hdbscan>=0.8.41",
    "librosa>=0.11.0",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pyannoteai-sdk>=0.4.0",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "sarvamai>=0.1.27",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "whisperx>=3.8.2",
]
